BuildCache

BinaryBuilder2.BuildCache — Type

BuildCache

This cache stores all necessary information to skip previously successful builds. If a previous build and extraction generated an artifact, stored its logs, analyzed the library dependency structure, etc... all of that is stored within the BuildCache so that future builds of the exact same source can be skipped. The cache keys are purposefully quite sensitive (if anything in the sources, or even just the sources of BinaryBuilder2 changes, the build will be re-run).

The BuildCache is serialized to disk and loaded on future BinaryBuilder2 sessions. Future work includes creating a networked backend to allow sharing of build cache objects across machines (or even the internet).

To disable the BuildCache, see the --disable-caches argument to build_tarballs.jl, or the disabled_caches keyword argument to BuildMeta.

Type Usage

The identifying hash of a BuildConfig is defined by spec_hash(::BuildConfig), which is sensitive to:

BuildTargetSpec objects, representing all compilers.

BuildConfig source trees, containing all dependencies.

BuildConfig script.

BinaryBuilder2 and dependent module source hashes.

To see this in action, you can run spec_hash(build_config) with a logger set to debug, and it will print out both the hash and all the inputs that went into the hash. Note that you must call prepare(build_config) first, to instantiate all dependencies.

julia> debug_logger = ConsoleLogger(stderr, Logging.Debug)
       with_logger(debug_logger) do
           spec_hash(build_config; force_recompute=true)
       end
┌ Debug: BuildConfig hash buffer:
│ [build_metadata]
│   script_hash = sha1:5ffe533b830f08a0326348a9160afafc8ada44db
│ [target_specs]
│   host: x86_64-linux-gnu-target_libc+glibc-target_os+linux-target_arch+x86_64
│   target: x86_64-linux-gnu-target_libc+glibc-target_os+linux-target_arch+x86_64
│ [source_trees]
│   /opt/host-tools = sha1:a57234ff529dc3fa30382935abee0a5d8007fe65
│   /opt/host-x86_64-linux-gnu = sha1:03d3df1856c3b5780665ed6a787a748d03ab81e0
│   /opt/target-x86_64-linux-gnu = sha1:da1cc815fcc5b8f7e79311c505e297e9e132d889
│   /usr/local = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│   /usr/share/licenses = sha1:d2d5d79765fa352588f9fc23b2abfef1a5d0fb96
│   /workspace/destdir/target-x86_64-linux-gnu = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│   /workspace/metadir = sha1:084f94c9181f34e7466e53eae6bfffcec61903eb
│   /workspace/scripts = sha1:a7189afd9cd7aa875783487b074364ffa0a276ed
│   /workspace/srcdir = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│ [environment]
│   BinaryBuilder2 = sha1:aac51e4d4aba12ae6618aa804cda9d68125076bf
│   BinaryBuilderAuditor = sha1:a1cb1172a8fa24028feba7361d6950a3fa2fa7d8
│   BinaryBuilderGitUtils = sha1:3ed475af33b46cdd85519443096409c850286e4b
│   BinaryBuilderPlatformExtensions = sha1:88784383a64abcf571e47ed19a0a97349aaf6564
│   BinaryBuilderProducts = sha1:5fa7cd339ed1f8d5b8d074512da6569354ff9ac2
│   BinaryBuilderSources = sha1:6afda911aa3216dd9fb42f2c1f38c14e1c0590dc
│   BinaryBuilderToolchains = sha1:ed992a365101e14803f3615af90a75d19fdf8c00
│   JLLGenerator = sha1:cb958a5e87299c4789cdeab4573f2e2f4706bd80
│   JLLPrefixes = sha1:6e132d90cd5e9bdb8025908e7a7971d72c0defa3
│   LazyJLLWrappers = sha1:063a7e939eb688406e360542f12e73828fea55f2
│   MultiHashParsing = sha1:f5d0a4ceb55dff17e345b169ca9564cc173afede
│   Sandbox = sha1:c7156ad981aa1501a83f5a140f08ee30692c3674
│   TreeArchival = sha1:7ea07d94f984690a10879d06491661a1270423bb
└ @ BinaryBuilder2 ~/src/BB2/src/build_api/BuildConfig.jl:287
sha1:b5c29c56ec883b47fda84b0a960c742e591bfbdf

The same can be done for an ExtractConfig object (note that it prints out the BuildConfig object as well during debugging, because it hashes that as an input as well!)

julia> debug_logger = ConsoleLogger(stderr, Logging.Debug)
       with_logger(debug_logger) do
           spec_hash(extract_config; force_recompute=true)
       end
┌ Debug: BuildConfig hash buffer:
│ [build_metadata]
│   script_hash = sha1:5ffe533b830f08a0326348a9160afafc8ada44db
│ [target_specs]
│   host: x86_64-linux-gnu-target_libc+glibc-target_os+linux-target_arch+x86_64
│   target: x86_64-linux-gnu-target_libc+glibc-target_os+linux-target_arch+x86_64
│ [source_trees]
│   /opt/host-tools = sha1:a57234ff529dc3fa30382935abee0a5d8007fe65
│   /opt/host-x86_64-linux-gnu = sha1:03d3df1856c3b5780665ed6a787a748d03ab81e0
│   /opt/target-x86_64-linux-gnu = sha1:a239c3ea36916f77024d49ee0797232dcc0578c8
│   /usr/local = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│   /usr/share/licenses = sha1:d2d5d79765fa352588f9fc23b2abfef1a5d0fb96
│   /workspace/destdir/target-x86_64-linux-gnu = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│   /workspace/metadir = sha1:084f94c9181f34e7466e53eae6bfffcec61903eb
│   /workspace/scripts = sha1:a7189afd9cd7aa875783487b074364ffa0a276ed
│   /workspace/srcdir = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│ [environment]
│   BinaryBuilder2 = sha1:aac51e4d4aba12ae6618aa804cda9d68125076bf
│   BinaryBuilderAuditor = sha1:a1cb1172a8fa24028feba7361d6950a3fa2fa7d8
│   BinaryBuilderGitUtils = sha1:3ed475af33b46cdd85519443096409c850286e4b
│   BinaryBuilderPlatformExtensions = sha1:88784383a64abcf571e47ed19a0a97349aaf6564
│   BinaryBuilderProducts = sha1:5fa7cd339ed1f8d5b8d074512da6569354ff9ac2
│   BinaryBuilderSources = sha1:6afda911aa3216dd9fb42f2c1f38c14e1c0590dc
│   BinaryBuilderToolchains = sha1:ed992a365101e14803f3615af90a75d19fdf8c00
│   JLLGenerator = sha1:cb958a5e87299c4789cdeab4573f2e2f4706bd80
│   JLLPrefixes = sha1:6e132d90cd5e9bdb8025908e7a7971d72c0defa3
│   LazyJLLWrappers = sha1:063a7e939eb688406e360542f12e73828fea55f2
│   MultiHashParsing = sha1:f5d0a4ceb55dff17e345b169ca9564cc173afede
│   Sandbox = sha1:c7156ad981aa1501a83f5a140f08ee30692c3674
│   TreeArchival = sha1:7ea07d94f984690a10879d06491661a1270423bb
└ @ BinaryBuilder2 ~/src/BB2/src/build_api/BuildConfig.jl:287
┌ Debug: ExtractConfig hash buffer:
│ [extraction_metadata]
│   build_hash = sha1:68ab520f17620647451edd4574d973d2c82e63d3
│   script_hash = sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709
│ [products]
│   libfoo = ["libfoo"]
└ @ BinaryBuilder2 ~/src/BB2/src/build_api/ExtractConfig.jl:92
sha1:e016baf0e743c75d85dc26f540eda8572117b83c

Future work

It would be nice to not load everything into memory at boot, and instead only look things up when necessary. Otherwise, I could imagine this getting slightly unwieldy in the future, for large BuildCaches.

We should come up with a good default heuristic for when to call prune!(bc).

When hacking on BB2 itself, the BuildCache is all but useless, because of the [environment] section of the spec_hash(::BuildConfig) changing every time. It would be nice if the BuildCache was either more granular (e.g. only depended on certain functions within BB2? Is such a thing possible?)