PackageConfig

BinaryBuilder2.PackageConfig — Type

PackageConfig

Inputs needed to determine how a set of extractions are packaged. Multiple extractions can be packaged together into the same JLL, for instance you can create a "library-only" variant of a JLL, and a variant with all headers and documentation included as well.

Type Usage